requirements.txt
discord.py==1.7.3
python-dotenv==1.0.0
discord-py-slash-command==3.0.3